* 'master' of https://github.com/awslabs/autogluon: (46 commits)
  [core] move transformers to setup_utils, bump dependency version (#3984)
  [AutoMM] Fix one lightning upgrade issue (#3991)
  [CI][Feature] Create a package version table (#3972)
  [v.1.1][Upgrade] PyTorch 2.1 and CUDA 12.1 upgrade (#3982)
  [WIP] Code implementation of Conv-LoRA (#3933)
  [timeseries] Ensure that all metrics handle missing values in the target (#3966)
  [timeseries] Fix path and device bugs (#3979)
  [AutoMM]Remove grounding-dino (#3974)
  [Docs] Update install modules content (#3976)
  Add note on pd.to_datetime (#3975)
  [AutoMM] Improve DINO performance (#3970)
  Minor correction in differ to pick correct environment (#3968)
  Fix windows python 3.11 issue by removing ray (#3956)
  [CI][Feature] Package Version Comparator (#3962)
  [timeseries] Add support for categorical covariates (#3874)
  [timeseries] Add method for plotting forecasts (#3889)
  Update conf.py copyright to reflect current year (#3932)
  [Timeseries][CI]Refactor CI to skip AutoMM and Tabular tests w.r.t timeseries changes (#3942)
  Fix HPO crash in memory check (#3931)
  [AutoMM][CI] Capping scikit-learn to avoid HPO test failure (#3947)
  ...

+++ timeseries/src/autogluon/timeseries/models/gluonts/abstract_gluonts.py
@@ -18,7 +18,6 @@ from gluonts.model.predictor import Predictor as GluonTSPredictor
 from pandas.tseries.frequencies import to_offset
 
 from autogluon.common.loaders import load_pkl
-from autogluon.common.utils.log_utils import set_logger_verbosity
 from autogluon.core.hpo.constants import RAY_BACKEND
 from autogluon.timeseries.dataset.ts_dataframe import ITEMID, TIMESTAMP, TimeSeriesDataFrame
 from autogluon.timeseries.models.abstract import AbstractTimeSeriesModel
@@ -205,7 +204,7 @@ class AbstractGluonTSModel(AbstractTimeSeriesModel):
             model = load_pkl.load(path=os.path.join(path, cls.model_file_name), verbose=verbose)
             if reset_paths:
                 model.set_contexts(path)
-            model.gts_predictor = PyTorchPredictor.deserialize(Path(path) / cls.gluonts_model_path)
+            model.gts_predictor = PyTorchPredictor.deserialize(Path(path) / cls.gluonts_model_path, device="auto")
         return model
 
     def _get_hpo_backend(self):
@@ -383,7 +382,6 @@ class AbstractGluonTSModel(AbstractTimeSeriesModel):
             if "lightning" in logger_name:
                 pl_logger = logging.getLogger(logger_name)
                 pl_logger.setLevel(logging.ERROR if verbosity <= 3 else logging.INFO)
-        set_logger_verbosity(verbosity, logger=logger)
         gts_logger.setLevel(logging.ERROR if verbosity <= 3 else logging.INFO)
 
         if verbosity > 3:
@@ -395,6 +393,7 @@ class AbstractGluonTSModel(AbstractTimeSeriesModel):
         self._check_fit_params()
         # update auxiliary parameters
         init_args = self._get_estimator_init_args()
+        keep_lightning_logs = init_args.pop("keep_lightning_logs", False)
         callbacks = self._get_callbacks(
             time_limit=time_limit,
             early_stopping_patience=None if val_data is None else init_args["early_stopping_patience"],
@@ -413,7 +412,7 @@ class AbstractGluonTSModel(AbstractTimeSeriesModel):
                 self.gts_predictor.batch_size = init_args["predict_batch_size"]
 
         lightning_logs_dir = Path(self.path) / "lightning_logs"
-        if lightning_logs_dir.exists() and lightning_logs_dir.is_dir():
+        if not keep_lightning_logs and lightning_logs_dir.exists() and lightning_logs_dir.is_dir():
             logger.debug(f"Removing lightning_logs directory {lightning_logs_dir}")
             shutil.rmtree(lightning_logs_dir)
